Preserve movie color when filtering
Switch data structures to allow for this (at least to be done right simply).
Instead of keeping an unfiltered list, a filtered list, and a mapping of indexes between them used for coloring,
just store the unfiltered list, and a list of indexes within it, which can be sorted/shuffled/filtered, even duplicated for wrapping if that's a helpful strategy, and implicitly has the information for coloring because it's a list of indexes into the unfiltered list.